Add platform build tags
