+++ pkg/envoy/resource_monitor_other.go
@@ -1,0 +1,9 @@
+//go:build !linux
+
+package envoy
+
+import "errors"
+
+func NewSharedResourceMonitor(tempDir string) (ResourceMonitor, error) {
+	return nil, errors.New("unsupported platform")
+}
